requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
cmdstanpy==1.1.0
contourpy==1.1.0
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.36
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-bootstrap-v5==1.0.11
django-import-export==3.1.0
django-pandas==0.6.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
ephem==4.1.4
et-xmlfile==1.1.0
fonttools==4.40.0
holidays==0.27.1
importlib-resources==5.12.0
joblib==1.3.1
kiwisolver==1.4.4
LunarCalendar==0.0.9
MarkupPy==1.14
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.1
packaging==23.1
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
plotly==5.15.0
pmdarima==2.0.3
prophet==1.1.4
PyMeeus==0.5.12
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
statsmodels==0.14.0
tablib==3.3.0
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
